+++ test/e2e/sql/src/test/java/org/apache/shardingsphere/test/e2e/engine/type/dql/GeneralDQLE2EIT.java
@@ -104,8 +104,8 @@ class GeneralDQLE2EIT extends BaseDQLE2EIT {
     
     private void assertExecuteQueryWithExpectedDataSource(final AssertionTestParameter testParam, final E2ETestContext context) throws SQLException {
         try (
-                Connection actualConnection = getEnvironmentEngine().getTargetDataSource().getConnection();
-                Connection expectedConnection = getExpectedDataSource().getConnection()) {
+                Connection expectedConnection = getExpectedDataSource().getConnection();
+                Connection actualConnection = getEnvironmentEngine().getTargetDataSource().getConnection()) {
             if (SQLExecuteType.LITERAL == context.getSqlExecuteType()) {
                 assertExecuteQueryForStatement(context, actualConnection, expectedConnection, testParam);
             } else {
@@ -117,10 +117,10 @@ class GeneralDQLE2EIT extends BaseDQLE2EIT {
     private void assertExecuteQueryForStatement(final E2ETestContext context, final Connection actualConnection, final Connection expectedConnection,
                                                 final AssertionTestParameter testParam) throws SQLException {
         try (
-                Statement actualStatement = actualConnection.createStatement();
-                ResultSet actualResultSet = actualStatement.executeQuery(context.getSQL());
                 Statement expectedStatement = expectedConnection.createStatement();
-                ResultSet expectedResultSet = expectedStatement.executeQuery(context.getSQL())) {
+                ResultSet expectedResultSet = expectedStatement.executeQuery(context.getSQL());
+                Statement actualStatement = actualConnection.createStatement();
+                ResultSet actualResultSet = actualStatement.executeQuery(context.getSQL())) {
             assertResultSet(actualResultSet, expectedResultSet, testParam);
         }
     }
@@ -128,15 +128,15 @@ class GeneralDQLE2EIT extends BaseDQLE2EIT {
     private void assertExecuteQueryForPreparedStatement(final E2ETestContext context, final Connection actualConnection, final Connection expectedConnection,
                                                         final AssertionTestParameter testParam) throws SQLException {
         try (
-                PreparedStatement actualPreparedStatement = actualConnection.prepareStatement(context.getSQL());
-                PreparedStatement expectedPreparedStatement = expectedConnection.prepareStatement(context.getSQL())) {
+                PreparedStatement expectedPreparedStatement = expectedConnection.prepareStatement(context.getSQL());
+                PreparedStatement actualPreparedStatement = actualConnection.prepareStatement(context.getSQL())) {
             for (SQLValue each : context.getAssertion().getSQLValues()) {
                 actualPreparedStatement.setObject(each.getIndex(), each.getValue());
                 expectedPreparedStatement.setObject(each.getIndex(), each.getValue());
             }
             try (
-                    ResultSet actualResultSet = actualPreparedStatement.executeQuery();
-                    ResultSet expectedResultSet = expectedPreparedStatement.executeQuery()) {
+                    ResultSet expectedResultSet = expectedPreparedStatement.executeQuery();
+                    ResultSet actualResultSet = actualPreparedStatement.executeQuery()) {
                 assertResultSet(actualResultSet, expectedResultSet, testParam);
             }
         }
@@ -220,7 +220,7 @@ class GeneralDQLE2EIT extends BaseDQLE2EIT {
         try (
                 Statement actualStatement = actualConnection.createStatement();
                 Statement expectedStatement = expectedConnection.createStatement()) {
-            assertTrue(actualStatement.execute(context.getSQL()) && expectedStatement.execute(context.getSQL()), "Not a query statement.");
+            assertTrue(expectedStatement.execute(context.getSQL()) && actualStatement.execute(context.getSQL()), "Not a query statement.");
             try (
                     ResultSet actualResultSet = actualStatement.getResultSet();
                     ResultSet expectedResultSet = expectedStatement.getResultSet()) {
@@ -238,7 +238,7 @@ class GeneralDQLE2EIT extends BaseDQLE2EIT {
                 actualPreparedStatement.setObject(each.getIndex(), each.getValue());
                 expectedPreparedStatement.setObject(each.getIndex(), each.getValue());
             }
-            assertTrue(actualPreparedStatement.execute() && expectedPreparedStatement.execute(), "Not a query statement.");
+            assertTrue(expectedPreparedStatement.execute() && actualPreparedStatement.execute(), "Not a query statement.");
             try (
                     ResultSet actualResultSet = actualPreparedStatement.getResultSet();
                     ResultSet expectedResultSet = expectedPreparedStatement.getResultSet()) {